+++ app/utils/constants.js
@@ -1,0 +1,15 @@
+const drawers = [
+  { stage: '0', queryInterval: 0 },
+  { stage: '1', queryInterval: 1 },
+  { stage: '2', queryInterval: 2 },
+  { stage: '3', queryInterval: 3 },
+  { stage: '4', queryInterval: 5 },
+  { stage: '5', queryInterval: 10 },
+  { stage: '6', queryInterval: 30 },
+  { stage: '7', queryInterval: 60 },
+  { stage: '8', queryInterval: 90 },
+];
+
+module.exports = {
+  drawers,
+};
